chore: rename LoCha to LoChaItem
